+++ src/utilities/math_lut.hh
@@ -1,3 +1,3 @@
 #include "lut.hh"
 
-extern Mapping::LookupTable_t<64, double> Pow2;
+extern Mapping::LookupTable_t<64, float> Pow2;
